+++ examples/nidcpower_source_dc_voltage/_helpers.py
@@ -3,7 +3,7 @@
 import logging
 import pathlib
 import types
-from typing import Any, Callable, Dict, NamedTuple, TypeVar, Optional
+from typing import Any, Callable, Dict, Iterable, NamedTuple, Optional, TypeVar, Tuple, Union
 
 import click
 import grpc
@@ -265,3 +265,39 @@ def create_session_management_client(
             service_class=nims.session_management.GRPC_SERVICE_CLASS,
         )
     )
+
+
+def get_session_and_channel_for_pin(
+    session_info: Iterable[nims.session_management.SessionInformation],
+    pin: str,
+    site: Optional[int] = None,
+) -> Tuple[int, Iterable[str]]:
+    """Returns the session information based on the given pin names."""
+    session_and_channel_info = get_sessions_and_channels_for_pins(pins=[pin], site=site)
+
+    if len(session_and_channel_info) != 1:
+        raise ValueError(f"Unsupported number of sessions for {pin}: {len(session_info)}")
+    return session_and_channel_info[0]
+
+
+def get_sessions_and_channels_for_pins(
+    session_info: Iterable[nims.session_management.SessionInformation],
+    pins: Union[str, Iterable[str]],
+    site: Optional[int] = None,
+) -> Iterable[Tuple[int, Iterable[str]]]:
+    """Returns the session information based on the given pin names."""
+    pin_names = [pins] if isinstance(pins, str) else pins
+    session_and_channel_info = []
+    for session_index, session_info in enumerate(session_info):
+        channel_list = [
+            mapping.channel
+            for mapping in session_info.channel_mappings
+            if mapping.pin_or_relay_name in pin_names and (site is None or mapping.site == site)
+        ]
+        if len(channel_list) != 0:
+            session_and_channel_info.append((session_index, channel_list))
+
+    if len(session_and_channel_info) == 0:
+        raise KeyError(f"Pin(s) {pins} and site {site} not found")
+
+    return session_and_channel_info
